Finally fix folder structure and file names, pfew. Also remove the data from repository, at least temporarily
