+++ src/main.py
@@ -1,10 +1,11 @@
 import asyncio
 import json
+import sys
 
 import aiohttp
 
-import CSVWriting
-import GetTwitchData
+import analysis
+import twitch
 
 
 async def main():
@@ -14,15 +15,18 @@ async def main():
         cr = json.load(open('../credentials.json'))
 
         # Get the top 100 streamers on Twitch
-        streamers = await GetTwitchData.get_top_streamers(session, cr, count=100)
+        streamers = await twitch.get_top_streamers(session, cr, count=100)
 
         # Create a dictionary of {streamer:[viewers]} from those 100 streams
-        data = await GetTwitchData.get_viewer_map(session, streamers)
+        data = await twitch.get_viewer_map(session, streamers)
 
         # Add that dictionary to the master csv
-        CSVWriting.update_twitch_data(data)
+        analysis.update_twitch_data(data)
 
 
 # Program Execution
 if __name__ == '__main__':
-    asyncio.run(main())
+    if len(sys.argv) > 1 and sys.argv[1] == 'export':
+        analysis.export()
+    else:
+        asyncio.run(main())
